pyproject.toml
[project]
name = "real-time-clean"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyaudio>=0.2.14",
    "google-cloud-speech>=2.26.0",
    "google-genai>=1.0.0",
    "asciimatics>=1.15.0",
    "instructor>=1.11.3",
    "jsonref>=1.1.0",
]
